feature/69 refactor: alias 적용
