+++ front/src/components/DoranAppBar.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import router from "../router/index";
+import router from "@/router";
 
 export default {
   name: "DoranAppBar",
