+++ front/src/components/PostItem.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script>
-import router from "../router";
+import router from "@/router";
 
 export default {
   name: "PostItem",
