+++ front/src/plugins/kakao-map.js
@@ -1,7 +1,7 @@
-import { KAKAO_MAP_APP_KEY } from "../secure/appkey";
-import INITIAL_LOCATION from "../config/config";
-import { ERROR_MESSAGE } from "../utils/constants";
-import { POST_OVERLAY_TEMPLATES } from "../utils/template";
+import { KAKAO_MAP_APP_KEY } from "@/secure/appkey";
+import INITIAL_LOCATION from "@/config/config";
+import { ERROR_MESSAGE } from "@/utils/constants";
+import { POST_OVERLAY_TEMPLATES } from "@/utils/template";
 
 const KAKAO_MAP_URL = "//dapi.kakao.com/v2/maps/sdk.js?autoload=false&appkey=";
 const LIBRARY = "&libraries=services";
