+++ front/src/router/index.js
@@ -1,16 +1,16 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
-import MainPage from "../components/MainPage";
-import PostDetailPage from "../components/PostDetailPage";
-import TimelinePage from "../components/TimelinePage";
+import MapPage from "@/components/map/MapPage";
+import PostDetailPage from "@/components/post/PostDetailPage";
+import TimelinePage from "@/components/TimelinePage";
 
 Vue.use(VueRouter);
 
 const routes = [
   {
     path: "/",
-    name: "MainPage",
-    component: MainPage,
+    name: "MapPage",
+    component: MapPage,
   },
   {
     path: "/timeline",
